predict.bs {splines}	R Documentation

Evaluate a Spline Basis

Description

Evaluate a predefined spline basis at given values.

Usage

## S3 method for class 'bs':
predict(object, newx, ...)

## S3 method for class 'ns':
predict(object, newx, ...)

Arguments

object 	 the result of a call to bs or ns having attributes describing knots, degree, etc.

newx 	 the x values at which evaluations are required.

... 	 Optional additional arguments. At present no additional arguments are used.

Value

An object just like object, except evaluated at the new values of x.
These are methods for the generic function predict for objects inheriting from classes "bs" or "ns". See predict for the general behavior of this function.

See Also

bs, ns, poly.

Examples

require(stats)
basis <- ns(women$height, df = 5)
newX <- seq(58, 72, length.out = 51)
# evaluate the basis at the new data
predict(basis, newX)

[Package splines version 2.9.2 Index]