splineOrder {splines}	R Documentation

Determine the Order of a Spline

Description

Return the order of a spline object.

Usage

splineOrder(object)

Arguments

object 	 An object that inherits from class "spline".

Details

The order of a spline is the number of coefficients in each piece of the piecewise polynomial representation. Thus a cubic spline has order 4.

Value

A positive integer.

Author(s)

Douglas Bates and Bill Venables

See Also

splineKnots, interpSpline, periodicSpline

Examples

splineOrder( interpSpline( weight ~ height, women ) )

[Package splines version 2.9.2 Index]