splines-package {splines}	R Documentation

Regression Spline Functions and Classes

Description

Regression spline functions and classes.

Details

This package provides functions for working with regression splines using the B-spline basis, bs, and the natural cubic spline basis, ns.

For a complete list of functions, use library(help = "splines").

Author(s)

Douglas M. Bates bates@stat.wisc.edu and William N. Venables Bill.Venables@csiro.au

Maintainer: R Core Team R-core@r-project.org

[Package splines version 3.6.1 Index]