xyVector {splines}	R Documentation

Construct an xyVector Object

Description

Create an object to represent a set of x-y pairs. The resulting object can be treated as a matrix or as a data frame or as a vector. When treated as a vector it reduces to the y component only.

The result of functions such as predict.spline is returned as an xyVector object so the x-values used to generate the y-positions are retained, say for purposes of generating plots.

Usage

xyVector(x, y)

Arguments

x 	 a numeric vector

y 	 a numeric vector of the same length as x

Value

An object of class xyVector with components

x 	 a numeric vector

y 	 a numeric vector of the same length as x

Author(s)

Douglas Bates and Bill Venables

Examples

require(stats); require(graphics)
ispl <- interpSpline( weight ~ height, women )
weights <- predict( ispl, seq( 55, 75, length.out = 51 ))
class( weights )
plot( weights, type = "l", xlab = "height", ylab = "weight" )
points( women$height, women$weight )
weights

[Package splines version 2.9.2 Index]